New Update 1.2.0
New function added to the instance, config(), now it is possible to change in real time the custom configuration in an instance started using .config(options)
New global Loggings function, it is now possible to set the global settings for all loggings instances using LoggingsConfig(options) making it possible to shape config at any time in the code, remember that this changes the logging standards settings to all instances, but If the instance has any custom configuration that overlaps the global configuration, the instance will still work with its modified configuration.
New standard configuration "remove_colors", this can be useful for hosts that make the log terminal, read Type to learn more.